feat(societies): move TruncateDescriptionContainer to common folder
 - refactor societies path
 - add loggedActivities argument in fetchSocietyInfoSuccess action
