+++ src/app/Dashboard/components/index.js
@@ -1,4 +1,3 @@
 export { default } from './DashboardContainer';
 export { default as MyStatsComponent } from './MyStatsComponent';
 export { default as SocietyStatsComponent } from './SocietyStatsComponent';
-export { default as TruncateDescriptionComponent } from './TruncateDescriptionContainer';
