+++ src/app/Societies/operations/types.js
@@ -8,4 +8,4 @@ export default {
   SOCIETY_PAGE_LOADING,
   FETCH_SOCIETY_INFO_REQUEST,
   FETCH_SOCIETY_INFO_SUCCESS,
-}
+};
