+++ src/app/common/components/index.js
@@ -7,3 +7,4 @@ export { default as SidebarComponent } from './SidebarComponent';
 export { default as NavItemComponent } from './NavItemComponent';
 export { default as ProgressBarComponent } from './ProgressBarComponent';
 export { default as StatusIndicatorComponent } from './StatusIndicatorComponent';
+export { default as TruncateDescriptionContainer } from './TruncateDescriptionContainer';
