+++ src/app/common/components/tests/TruncateDescriptionContainer.test.js
